requirements.txt
Django==5.0.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.1
PyJWT==2.8.0
Pillow==11.2.1
psycopg2==2.9.9
celery==5.3.6
redis==5.0.1
drf-yasg==1.21.7
requests==2.31.0 